fixed routing bug which had caused only the first ChatRoom clicked on to render
